bringup: arm32bit on kernel6.12 [1/1]
PD#SWPL-188330

Problem:
arm32bit on kernel6.12

Solution:
1. modify build script
2. save defconfig for a32_base_defconfig
3. sync amlogic_gki.fragment to amlogic_a32.fragment
4. disable below configs:
CONFIG_AMLOGIC_FILE_CACHE
CONFIG_AMLOGIC_WIRELESS
CONFIG_AMLOGIC_MEMORY_EXTEND
CONFIG_AMLOGIC_PAGE_TRACE
CONFIG_AMLOGIC_SLAB_TRACE
CONFIG_AMLOGIC_CMA
CONFIG_AMLOGIC_VMAP

Verify:
local

+++ drivers/tty/serial/meson_uart.c
@@ -503,7 +503,7 @@ static int meson_uart_request_port(struct uart_port *port)
 			return -ENOMEM;
 	}
 
-	dev_dbg(&pdev->dev, "==uart%d reg addr = 0x%llx\n", port->line, port->mapbase);
+	dev_dbg(&pdev->dev, "==uart%d reg addr = 0x%llx\n", port->line, (unsigned long long)port->mapbase);
 
 	val = (AML_UART_RECV_IRQ(1) | AML_UART_XMIT_IRQ(port->fifosize / 2));
 	writel_relaxed(val, port->membase + AML_UART_MISC);
